requirements.txt
paddleocr>=2.6.0.3
paddleclas>=2.4.3
pytest-runner
paddlespeech==1.2.0
paddlenlp<=2.0.5
torch>1.7.0
Pillow==9.5.0
nltk
pystray
wave
pydub
pynput
pyperclip
ultralytics == 8.0.120
pandas>=1.1.4
tqdm>=4.64.0
scipy>=1.4.1
requests>=2.23.0
PyYAML>=5.3.1
opencv-python>=4.6.0
ttkthemes
python-pptx
transformers
sacremoses
moviepy
pysrt
